feat: refactor code to use Provider package for state management
The code has been refactored to use the Provider package for state management. The `subscriptionsPageModel` has been moved to a separate file `subscriptions_page_model.dart` and is now created using the `Provider` widget. The `SubscriptionsView` and `HomeAppBar` classes have been updated to use the `Provider` widget and access the `subscriptionsPageModel` and `subscriptionsModel` respectively. This allows for better separation of concerns and improved code organization.