feat: add support for default currency setting in the app
The changes introduce a new `SettingsModel` class that handles the default
currency for the app. This allows users to set their preferred currency for
transactions. The `defaultCurrency` property is now used in the `RecurringTab`
and `LifetimeTab` classes to determine the currency displayed in the UI. The
`updateDefaultCurrency` method allows users to update their default currency
preference.

resolves #65